netscope/requirements-dev.txt
# NETSCOPE Development Dependencies
-r requirements.txt
pytest>=7.0.0
pytest-flask>=1.2.0
ruff>=0.1.0
